requirements.txt
asttokens==2.0.5
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
compress-pickle==2.1.0
cryptography==35.0.0
cycler==0.11.0
executing==0.8.3
fasttext==0.9.2
Flask==2.0.2
fonttools==4.29.1
gensim==4.1.2
icecream==2.1.2
idna==3.3
importlib-metadata==4.11.2
itsdangerous==2.0.1
jellyfish==0.9.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
nltk==3.7
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pipe==1.6.0
play-scraper==0.6.0
pry.py==0.1.1
psutil==5.9.0
pybind11==2.9.2
pycparser==2.20
Pygments==2.11.2
pyparsing==3.0.7
pyphonetics==0.5.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyxDamerauLevenshtein==1.7.0
realpath==0.1.0
regex==2022.3.2
requests==2.27.1
requests-futures==1.0.0
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
tqdm==4.63.0
Unidecode==1.3.3
urllib3==1.26.7
Werkzeug==2.0.2
wordsegment==1.3.1
zipp==3.7.0
